Modified NATS test to not incur a slow consumer.
Many messaging systems are inbalanced and can send faster than they can
receive. NATS has the ability to do this and was reporting a slow
consumer state that was not being realized. Adding in the original
flush alleviated the problem, but at a huge performance cost for
throughput. This change balances the publishing rate with the subscriber
rate.